% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clscolor.R
\name{clscolor}
\alias{clscolor}
\alias{valcolor}
\title{Cluster Colors: a range of colors for clusters or values}
\usage{
clscolor(n, s = 1, v = 0.8, alpha = 0.5)

valcolor(x, n)
}
\arguments{
\item{n}{the number of distinct colors to be generated.}

\item{s}{the saturation of HSV color with a default value of 1.}

\item{v}{the value of HSV color with a default value of 0.8.}

\item{alpha}{the alpha transparency value for the colors, with a default
value of 1.}

\item{x}{the vector of values, or a range of values.}

\item{lowColor}{the color for the lowest value in the range of x}

\item{highColor}{the color for the highest value in the range of x}
}
\value{
a function that returns a color for each number between 1 to \code{n}, used
as the hue in the HSV color scheme.

a function that returns a color for a numerical value.
}
\description{
A function to produce color for a range of values between two colors
}
