% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterModel.R
\name{clusterModel}
\alias{clusterModel}
\alias{predict.ClusterClass}
\title{Cluster Class: A Generic Distance Based Classification Model}
\usage{
clusterModel(x, y, center = "mean", dist = "euclidean")

\method{predict}{ClusterClass}(object, newdata, type = "response")
}
\arguments{
\item{x}{training data, a numerical matrix}

\item{y}{classification for training data}

\item{center}{if \code{mean} (default), uses the mean of each cluster as its center.
if a numerical matrix, it is used as centers.}

\item{dist}{if \code{euclidean} (default), the euclidean distance is used to
compute the distance from the center in prediction step; otherwise, this is
a function to compute the distance in prediction step.}

\item{object}{a ClusterClass object}

\item{newdata}{test data to be used for prediction using the object model}

\item{type}{if \code{response} (default), the predicted classes are retrurned.
If \code{distance} is used, the distance of each row in newdata to all clusters
is returned}
}
\description{
Prediction function for ClusterClass models
}
