% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqMag.R
\name{freqMag.vector}
\alias{freqMag.vector}
\title{Computing frequency & magnitudes of an FFT}
\usage{
freqMeg.vector(fftx)

freqMeg.matrix(fftx)
}
\arguments{
\item{fftx}{a vector or matrix of fast-fourier transformation}
}
\value{
a list of three values: mag (magnitude), ang (angle in rad), and freq (frequencies)
}
\description{
Computing frequency & magnitudes of an FFT
}
