% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ij2k.R
\name{ij2k}
\alias{ij2k}
\title{Matrix position index (i,j) to Vector position index k}
\usage{
ij2k(i, j = NULL, nrow)
}
\arguments{
\item{i}{a vector of positions in the first dimension, or a bi-column matrix
containing the positions in two dimensions.}

\item{j}{the vector of positions in the second dimension, if \code{i} is not a bi-column matrix}

\item{nrow}{the number of rows in the matrix representation of data.}
}
\value{
a vector of linear position indexes equivalent to the bi-column matrix
(i, j), representing position indexes in a matrix with \code{nrow} rows.
}
\description{
Matrix position index (i,j) to Vector position index k
}
