% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2ij.R
\name{k2ij}
\alias{k2ij}
\title{Vector position index k to Matrix position index (i,j)}
\usage{
k2ij(nrow, k)
}
\arguments{
\item{nrow}{the number of rows in the matrix representation of data.}

\item{k}{the list of linear position indexes.}
}
\value{
a bi-column matrix containing (i, j), showing the equivalent position for
positions in k in a linear presentation of a matrix with \code{nrow} rows.
}
\description{
Vector position index k to Matrix position index (i,j)
}
