% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_test.R
\name{kfold_test}
\alias{kfold_test}
\title{K-Fold Test
Runs a k-fold test of a given model, or a model trained based on the data}
\usage{
kfold_test(
  x,
  y,
  k = 10,
  model,
  predictor = predict,
  modelparams = NULL,
  summarize = TRUE,
  ...
)
}
\arguments{
\item{x}{the dataset to be used for test and training}

\item{y}{the response variable, a vector.}

\item{k}{the number of fold the dataset will be broken to, with a default value
of 10.}

\item{model}{an object representing a trained model, or a function to be used for
training the data.}

\item{predictor}{a function to predict the response for the test data based on the
trained model. The default function is \code{predict}.}

\item{modelparams}{is a list of values to be passed into \code{model}, if it is a function.}

\item{summarize}{if TRUE, the default value, the mean and standard deviation of
the performance measures are returned; otherwise all performance measured are returned.}

\item{...}{the list of parameters to be passed to \code{predictor} function.}
}
\value{
A data.frame including the fold number and performance measures.
}
\description{
K-Fold Test
Runs a k-fold test of a given model, or a model trained based on the data
}
