% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskr.R
\name{maskr}
\alias{maskr}
\title{Adding random error to masks}
\usage{
maskr(
  x,
  tilt_err = 0.1,
  twist_err = 1,
  mask_err = 1,
  orientation_err = 5,
  az_cover = 360,
  clipLim = c(-Inf, +Inf),
  clipValNA = FALSE
)
}
\arguments{
\item{x}{a vector of a mask or a matrix containing multiple masks as rows.}

\item{tilt_err}{determining the SD of error added to tilt angle in degrees.
Its default value is 0.1 degrees.}

\item{twist_err}{is the SD of error added to twist angle in degrees. Its
default value is 1.}

\item{mask_err}{is the SD of error added to mask value. Its default value is 1.}

\item{orientation_err}{is the SD error added to the orientation angle, with a
default of 5 degree.}

\item{az_cover}{the azimuth coverage of mask, with a default value of 360.}

\item{clipLim}{a set of lower and upper bounds for masks with no limits as default.}

\item{clipValNA}{a logical value; TRUE means use NA when mask is clipped; default is FALSE}
}
\description{
Adding random error to masks
}
