% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2img.R
\name{mat2img}
\alias{mat2img}
\title{Turning a matrix to an image}
\usage{
mat2img(x, scale = TRUE, ...)
}
\arguments{
\item{x}{the matrix with numerical values to be used as colors}

\item{scale}{if TRUE, the values are scaled to 0 to 255}

\item{...}{additional parameters sent to \code{image()}.}
}
\value{
returns the scaled matrix and plots the image.
}
\description{
Turning a matrix to an image
}
